aider: Reduced font size, padding, and size of social icons in header for mobile breakpoints.
# Aider chat conversation:

ASSISTANT: To optimize the header for mobile breakpoints, we need to modify the `styles/app.sass` file. We will reduce the font size and padding of the title in the header and also reduce the size of the social icons for smaller screens.

Here are the *SEARCH/REPLACE* blocks:

styles/app.sass
```sass
<<<<<<< SEARCH
.header
  width: 100%
  padding: 0.8rem
  display: flex
  justify-content: space-between
  align-items: center

.header .title
  margin-left: 2rem

.header .social-icons
  margin-right: 2rem
=======
@media (max-width: 800px)
  .header
    width: 100%
    padding: 0.4rem
    display: flex
    justify-content: space-between
    align-items: center

  .header .title
    margin-left: 1rem
    font-size: 1.2rem

  .header .social-icons
    margin-right: 1rem
    .icon
      font-size: 1.2rem
>>>>>>> REPLACE
```

This change will reduce the padding of the header, the margin and font size of the title, and the size of the social icons when the screen width is 800px or less.