aider: Adjusted styles for the HTML, body, footer, and main container to anchor the footer to the bottom of the page.
# Aider chat conversation:

ASSISTANT: To anchor the footer to the bottom of the page in both mobile and desktop views, we need to make sure that the main content of the page takes up at least the remaining height of the viewport if the content is not enough to push the footer down. We can achieve this by using flexbox on the container that wraps the main content and the footer.

We will modify the `styles/app.sass` file to adjust the styles for the `html`, `body`, and the main container (which we will assume has a class of `.container` based on your `pages/index.js` file). If the main container does not have a class of `.container`, please replace `.container` with the appropriate class or element selector.

Here are the *SEARCH/REPLACE* blocks:

styles/app.sass
```sass
<<<<<<< SEARCH
html, body
  height: 100%
  margin: 0
  padding: 0
=======
html, body
  min-height: 100%
  margin: 0
  padding: 0
  display: flex
  flex-direction: column
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
=======
footer
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
  margin-top: auto // Pushes the footer to the bottom
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
main
  padding: 0 0.5rem
  overflow: hidden
  display: flex
  flex-direction: column
  justify-content: start
  align-items: center
=======
main
  padding: 0 0.5rem
  overflow: hidden
  display: flex
  flex-direction: column
  justify-content: start
  align-items: center
  flex: 1 0 auto // Allows the main content to grow and fill the available space
>>>>>>> REPLACE
```

Please ensure that your main container in `pages/index.js` has a class of `.container`. If it does not, you will need to add it or adjust the class name in the SASS file accordingly.