aider: Anchored the footer at the bottom of the page by modifying the `styles/app.sass` file to use CSS Flexbox and prevent the footer from shrinking.
# Aider chat conversation:

ASSISTANT: To anchor the footer at the bottom of the page, we need to modify the `styles/app.sass` file. We will use CSS Flexbox to make the container fill the height of the page and push the footer to the bottom.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
html, body
  height: 100%
  margin: 0
  padding: 0
=======
html, body, .container
  height: 100%
  margin: 0
  padding: 0
  display: flex
  flex-direction: column
>>>>>>> REPLACE
```

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
  margin-top: auto // Pushes the footer to the bottom
=======
footer
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
  margin-top: auto // Pushes the footer to the bottom
  flex-shrink: 0
>>>>>>> REPLACE
```

This change will make the `.container` div (which wraps all the content in your `index.js` file) fill the height of the page, and the `flex-shrink: 0` property will prevent the footer from shrinking, effectively anchoring it to the bottom of the page.