aider: Changed main content area to grow and fill the available space.
# Aider chat conversation:

USER: I think we also need to have main grow to be as big as possible right?

ASSISTANT: Yes, you're correct. To ensure that the main content area takes up as much space as possible, we can add `flex-grow: 1` to the main selector in your `styles/app.sass` file. This will allow the main content area to grow and fill the available space, pushing the footer to the bottom of the page.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
main
  padding: 0 0.5rem
  overflow: hidden
  display: flex
  flex-direction: column
  justify-content: flex-start
  align-items: center
  flex: 1 0 auto // Allows the main content to grow and fill the available space
=======
main
  padding: 0 0.5rem
  overflow: hidden
  display: flex
  flex-direction: column
  justify-content: flex-start
  align-items: center
  flex-grow: 1 // Allows the main content to grow and fill the available space
>>>>>>> REPLACE
```

This change will make the main content area grow to fill the available space, effectively pushing the footer to the bottom of the page.