aider: Changed footer position to absolute and set it to the bottom of the page.
# Aider chat conversation:

USER: lol this isn't working, why don't we just absolutely position the footer at the bottom

ASSISTANT: To absolutely position the footer at the bottom of the page, we need to modify the `styles/app.sass` file. We will set the position of the footer to absolute and place it at the bottom of the page.

Here is the *SEARCH/REPLACE* block:

styles/app.sass
```sass
<<<<<<< SEARCH
footer
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
  margin-top: auto // Pushes the footer to the bottom
  flex-shrink: 0
=======
footer
  position: absolute
  bottom: 0
  width: 100%
  height: 50px
  display: flex
  justify-content: center
  align-items: center
>>>>>>> REPLACE
```

This change will make the footer stick to the bottom of the page, regardless of the content above it.