aider: Set the background of the `.header` to the same as the `.card`.
# Aider chat conversation:

USER: Let's make the background of the header the same as the cards

ASSISTANT: To make the background of the `.header` the same as the `.card`, you will need to add a background property to the `.header` class in your `styles/app.sass` file. However, since the `.card` background color is defined in a separate file (`./themes/dark-theme.css`), you should ensure that the background color is consistent by using the same variable or class name.

If the background color for `.card` is a CSS variable or a SASS variable, you would use that. If it's a class, you would add that class to the `.header`. Since I don't have the `dark-theme.css` file, I'll show you how to add a placeholder for the background color, which you should replace with the actual color or variable used for the `.card`.

styles/app.sass
```sass
<<<<<<< SEARCH
.header
  width: 100%
  padding: 0.4rem
  display: flex
  justify-content: space-between
  align-items: center
=======
.header
  width: 100%
  padding: 0.4rem
  display: flex
  justify-content: space-between
  align-items: center
  background: $card-background-color // Replace with the actual variable or color used for .card
>>>>>>> REPLACE
```

Please replace `$card-background-color` with the actual variable or color code that you're using for the `.card` background in `./themes/dark-theme.css`. If you're using a CSS class instead, you would add that class to the `.header` in your HTML file.