aider: Add LinkedIn style circular headshot to the left of the title in the Header.
# Aider chat conversation:

ASSISTANT: To add a LinkedIn style circular headshot to the left of the title in the Header, we need to modify `components/Header.js`:

1. Import the Image component from 'next/image'.
2. Add the Image component before the title in the Header component.

Here are the *SEARCH/REPLACE* blocks:

components/Header.js
```javascript
<<<<<<< SEARCH
import Head from 'next/head'
=======
import Head from 'next/head'
import Image from 'next/image'
>>>>>>> REPLACE
```

components/Header.js
```javascript
<<<<<<< SEARCH
    <h1 className="title">
      SM Userspace {"</>"}
    </h1>
=======
    <div style={{ display: 'flex', alignItems: 'center' }}>
      <Image
        src="/images/headshot.jfif"
        alt="Headshot"
        width={50}
        height={50}
        className="rounded-full"
      />
      <h1 className="title">
        SM Userspace {"</>"}
      </h1>
    </div>
>>>>>>> REPLACE
```

Please note that the `width` and `height` attributes in the `Image` component are required for Next.js to determine the aspect ratio of the image. You may need to adjust these values to fit your needs. The `rounded-full` class is a Tailwind CSS utility class that makes the image circular. If you're not using Tailwind CSS, you'll need to replace this with the appropriate CSS to make the image circular.